feat(db): implement base db and assignment management functionality (#3)
* chore: begin creating content database services

Begin creating the SQLAlchemy logic to handle schema creation.

* chore: set up database tables for db features

Sets up the initial database tables for db features. Note that since
the `projects` table now has more relationships, the test suite fails -
this will be fixed in the next commit.

* chore: make progress on assignment creation flow

Makes changes to prepare for creating assignments and provisioning
necessary db schemas and roles.

* chore: set up db reset scripts, prep content db functionality

This commit sets up the database reset scripts and some testing.
The multi-schema multi-tenant database model was tested, but it is looking
like Tinkerbase will move to a mulit-db model for proper isolation.

* chore: working proof-of-concept multi db setup

Working proof-of-concept for a multi-database setup for the multi-
tenant database structure.

* chore: fix role creation and testing db setup

This commit ensures that role creation works successfully for both
admin and read-only (view) users, as well as fixes the `conftest` setup
so that databases and roles in the content cluster are properly dropped.

* chore: refactor to use new `content_db_cluster` service

Refactors the codebase to utilize the new `content_db_cluster` service
instead of the defunct `content_database` service, and removed the
`content_database` service entirely.

* chore: implement the course service

Complete the first implementation of the course service. Testing still
has to be completed.

* chore: set up test seeding, complete course svc tests

Set up the infrastructure for test seeding and model structure, as well
as completed the course service tests with nearly full coverage.

* chore: start work on the assignment service

Incremental commit to get started on the assignment service and testing.

* chore: create assignment service rename functionality

Create and fully test the assignment service's rename capability.

* chore: add support for adding config sql to assignments

Adds the service functionality and testing for adding config SQL to
assignments, as well as the ability to reset databases.

* chore: add save sql config to assignments

Completed the SQL configuration feature by adding the ability to save
the SQL configuration after testing.

* chore: add assignment group management

This commit adds CRUD operations for managing assignment / project
groups.

* chore: implement assignment publishing capability, add redis

This commit adds the ability for assignments to be published. A redis
instance has also been added to the project to handle task queuing
via the Celery package.

* chore: implement deleting, unpublising, and republishing assignments

This commit adds functionality to delete, unpublish, and republish
assignments.