feat(web): implement web app for base functionality (#4)
* chore: add theme to web app

Adds the theme and light/dark mode support to the web application.

* chore: configure fonts and base layout

Configures the theme's fonts to load correctly and set up the base
layout for the landing page.

* chore: begin work on the login page

Begins work on the UI for the login page.

* chore: implement web app authentication

This commit adds authentication for the web app, including the
necessary backend API endpoints and frontend components. This commit
also adds a helper for protecting frontend pages for only
authenticated users.

* chore: set up web app, add dashboard and login management features

Sets up many base features of the web application, including the login
management and related APIs. Also adds the dashboard page and related
APIs.

This commit sets up the frontend to use schemas generated by OpenAPI,
which can be generated from within the `/web` directory by running
`npm run generate-types`.

* chore: set up course page sidebar and header

Sets up the course page's sidebar and header, including necessary
APIs for the header to function.

* chore: work on loading assignments onto course page

This commit adds the necessary APIs and frontend components to show
assignments on the course page.

* chore: make significant progress for assignments draft page

Make significant backend and frontend progress for working on managing
assignments.

* chore: add sql configuration dialog for assignment setup

This commit adds the SQL configuration dialog and necessary APIs
for assignment setup.

* chore: complete assignment creation (draft editing) and publishing

Completes the core functionality for assignment creation, including
publishing. This includes creating the necessary APIs, as well as
changing honcho to run the Celery process.

* chore: make significant progress on assignment staff view

Makes significant process on the assignment staff view, including the
paginated tables for viewing created projects for an assignment (and
all necessary APIs)

* chore: implement export all student db sql dumps

This commit implements all student databases as SQL dumps in a project.

* chore: implement unpublish, republish, and delete assignment functionality

Implements the ability to unpublish, republish, and delete assignments.

* chore: add contact option to project table

Add the ability for the instructor to contact students directly
from the list of projects.

* chore: add group editing from the staff view

Add the ability to edit group members from the staff view of the
assignment page.

* chore: add error state for projects in assignment staff page

Add error state handling for projects in the assignment staff page.

* chore: refactor react query calls to custom hooks

Create custom hooks for all queries / mutations for better abstraction
and cleaner component code.

* chore: create base student view scaffold

Sets up the base student view and related APIs so that students can
view their assignment database URL and public auth token.

* chore: implement course roster

Implements the course roster page for the staff view for a course.

* chore: implement course staff settings view

Adds a basic course settings view that allows editing of the course
code and course name.

Also disables the course code validation for no spaces (temporary
for now but might become permanent)

* chore: implement join course functionality

Allow students to join courses and for instructors to share their
course codes.

* chore: update landing page and add logo

Updates the landing page and adds the logo for better branding. Note
that all changes here are a work in progress - both the landing page
and the logo need to change.