feat(realtime): implement base realtime functionality (#6)
* chore: update Dockerfile to include elixir

Updated the Dockerfile to download Erlang and Elixir dependencies.

* chore: set up elixir / phoenix realtime project

Sets up the Elixir and Phoenix realtime project with a lot of the
skeleton code set up for setting up the websocket connection.

* chore: notify realtime server of project creation

Adds a notification to the realtime server when a new project is created.

* chore: configure realtime notifs for db, add realtime task for celery

Adds a realtime task to celery that polls every 5s for changes in the
tables of a database. If tables are added, the task will add the realtime
trigger to the new table.

* chore: configure reverse proxy, working pg_changes

Gets the websocket endpoint to work for pg_changes by configuring
the Caddy reverse proxy.

* chore: implement all realtime features

Implements all realtime features that minimally works.

* chore: ensure realtime db connections uses env variables

Sets up environment variables in the Elixir server.

* chore: remove unused websocket page

Removes the unused websocket testing page.