chore: implement setup / teardown db cluster scripts (#20)
* chore: implement configuration functions for db clusters

Adds setup and teardown functions for the configuration of the db
clusters.

* chore: create setup scripts

Creates the setup scripts for both the development and production
environments.

* chore: refactor testing

Refactors the testing to work with the new setup for scripts.

* chore: add teardown prod scripts

Add a script to teardown prod.

* chore: remove unused imports

* chore: configure alembic

Configures Alembic for admin database migrations.

* chore: fix alembic migrations path

Fixes the Alembic migrations path to include the server.

* chore: fix migrations path pt 2

Removes a leading slash in the path which causes Alembic to not
recognize the migration directory.