feat(admin): add admin functionality (#21)
* chore: add role and permission tables

Adds a table to store roles and permissions.

* chore: add permission service

Add the permission service and modify the permission logic to just
include permissions and users (removes roles).

* chore: add instructor flag to users

Adds a flag to determine whether or not a user is an instuctor.

* chore: ensure that only instructors can add courses

Add validation so that courses can only be created by instructors.

* chore: add set_admin script

Add a script that elevates a user to have admin permissions based on
their ID.

* feat: add admin frontend for instructor crud

Adds the ability for the admin user to access an admin portal on the website and modify instructor permissions.

* feat: add ui for creating courses

Add UI for creating courses as an instructor.

* chore: add env configuration to the web app

Adds an env configuration to the web app so that the host is known.