feat(storage): implement cloud storage functionality  (#30)
* build: add minio to dev environment

Adds minio to the development environment to model S3 cloud storage.

* refactor: rename realtime token to project token

Renames the realtime token to the project token so that its naming
is more generalized - the project token will also be used to
authenticate the storage endpoints.

* chore: regenerate frontend types based on refactor

Regenerates the frontend types based on the refactor mentioned
in the previous commit.

* chore: implement storage endpoints

Adds the API endpoints for storage functionality.